+++ tests/sksl/metal/CastMat2x3ToMat4x4.metal
@@ -4,7 +4,7 @@ using namespace metal;
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 float4x4 float4x4_from_float2x3(float2x3 x0) {
     return float4x4(float4(x0[0].xyz, 0.0), float4(x0[1].xyz, 0.0), float4(0.0, 0.0, 1.0, 0.0), float4(0.0, 0.0, 0.0, 1.0));
@@ -14,6 +14,6 @@ fragment Outputs fragmentMain(Inputs _in [[stage_in]], bool _frontFacing [[front
     (void)_out;
     float4x4 a = float4x4(6.0);
     float4x4 b = float4x4_from_float2x3(float2x3(7.0));
-    _out.sk_FragColor.x = float(all(a[1] == b[1]) ? 0 : 1);
+    _out.sk_FragColor.x = half(all(a[1] == b[1]) ? 0 : 1);
     return _out;
 }
