+++ tests/sksl/metal/SamplerGlobals.metal
@@ -4,15 +4,15 @@ using namespace metal;
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 struct Globals {
-    texture2d<float> texA;
+    texture2d<half> texA;
     sampler texASmplr;
-    texture2d<float> texB;
+    texture2d<half> texB;
     sampler texBSmplr;
 };
-fragment Outputs fragmentMain(Inputs _in [[stage_in]], texture2d<float> texA[[texture(1)]], sampler texASmplr[[sampler(1)]], texture2d<float> texB[[texture(0)]], sampler texBSmplr[[sampler(0)]], bool _frontFacing [[front_facing]], float4 _fragCoord [[position]]) {
+fragment Outputs fragmentMain(Inputs _in [[stage_in]], texture2d<half> texA[[texture(1)]], sampler texASmplr[[sampler(1)]], texture2d<half> texB[[texture(0)]], sampler texBSmplr[[sampler(0)]], bool _frontFacing [[front_facing]], float4 _fragCoord [[position]]) {
     Globals _globals{texA, texASmplr, texB, texBSmplr};
     (void)_globals;
     Outputs _out;
