+++ tests/sksl/shared/HexUnsigned.metal
@@ -2,12 +2,12 @@
 #include <simd/simd.h>
 using namespace metal;
 struct Uniforms {
-    float4 colorGreen;
+    half4 colorGreen;
 };
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 fragment Outputs fragmentMain(Inputs _in [[stage_in]], constant Uniforms& _uniforms [[buffer(0)]], bool _frontFacing [[front_facing]], float4 _fragCoord [[position]]) {
     Outputs _out;
@@ -20,7 +20,7 @@ fragment Outputs fragmentMain(Inputs _in [[stage_in]], constant Uniforms& _unifo
     u3++;
     uint u4 = 4294967294u;
     u4++;
-    uint u5 = 65534u;
+    ushort u5 = 65534u;
     u5++;
     _out.sk_FragColor = _uniforms.colorGreen;
     return _out;
