+++ tests/sksl/shared/StaticIf.metal
@@ -2,18 +2,18 @@
 #include <simd/simd.h>
 using namespace metal;
 struct Uniforms {
-    float4 colorRed;
-    float4 colorGreen;
+    half4 colorRed;
+    half4 colorGreen;
 };
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 fragment Outputs fragmentMain(Inputs _in [[stage_in]], constant Uniforms& _uniforms [[buffer(0)]], bool _frontFacing [[front_facing]], float4 _fragCoord [[position]]) {
     Outputs _out;
     (void)_out;
-    float4 result = _uniforms.colorRed;
+    half4 result = _uniforms.colorRed;
     {
         result = _uniforms.colorGreen;
     }
