+++ tests/sksl/shared/SwizzleScalarBool.metal
@@ -2,12 +2,12 @@
 #include <simd/simd.h>
 using namespace metal;
 struct Uniforms {
-    float unknownInput;
+    half unknownInput;
 };
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 fragment Outputs fragmentMain(Inputs _in [[stage_in]], constant Uniforms& _uniforms [[buffer(0)]], bool _frontFacing [[front_facing]], float4 _fragCoord [[position]]) {
     Outputs _out;
@@ -17,6 +17,6 @@ fragment Outputs fragmentMain(Inputs _in [[stage_in]], constant Uniforms& _unifo
     b4 = bool4(bool2(b), false, true);
     b4 = bool4(false, b, true, false);
     b4 = bool4(false, b, false, b);
-    _out.sk_FragColor = float4(b4);
+    _out.sk_FragColor = half4(b4);
     return _out;
 }
