+++ tests/sksl/shared/SwizzleScalarInt.metal
@@ -2,12 +2,12 @@
 #include <simd/simd.h>
 using namespace metal;
 struct Uniforms {
-    float unknownInput;
+    half unknownInput;
 };
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 fragment Outputs fragmentMain(Inputs _in [[stage_in]], constant Uniforms& _uniforms [[buffer(0)]], bool _frontFacing [[front_facing]], float4 _fragCoord [[position]]) {
     Outputs _out;
@@ -17,6 +17,6 @@ fragment Outputs fragmentMain(Inputs _in [[stage_in]], constant Uniforms& _unifo
     i4 = int4(int2(i), 0, 1);
     i4 = int4(0, i, 1, 0);
     i4 = int4(0, i, 0, i);
-    _out.sk_FragColor = float4(i4);
+    _out.sk_FragColor = half4(i4);
     return _out;
 }
